+++ mock/mock.go
@@ -11,8 +11,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/stretchr/objx"
-
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/internal/difflib"
 	"github.com/stretchr/testify/internal/spew"
@@ -311,7 +309,7 @@ type Mock struct {
 
 	// TestData holds any data that might be useful for testing.  Testify ignores
 	// this data completely allowing you to do whatever you like with it.
-	testData objx.Map
+	testData testData
 
 	mutex sync.Mutex
 }
@@ -322,16 +320,6 @@ type Mock struct {
 // without acquiring the mutex, which is detected by go test -race.
 func (m *Mock) String() string {
 	return fmt.Sprintf("%[1]T<%[1]p>", m)
-}
-
-// TestData holds any data that might be useful for testing.  Testify ignores
-// this data completely allowing you to do whatever you like with it.
-func (m *Mock) TestData() objx.Map {
-	if m.testData == nil {
-		m.testData = make(objx.Map)
-	}
-
-	return m.testData
 }
 
 /*
