+++ mock/mock_test.go
@@ -164,18 +164,6 @@ func (m *MockTestingT) FailNow() {
 	Mock
 */
 
-func Test_Mock_TestData(t *testing.T) {
-	t.Parallel()
-
-	var mockedService = new(TestExampleImplementation)
-
-	if assert.NotNil(t, mockedService.TestData()) {
-
-		mockedService.TestData().Set("something", 123)
-		assert.Equal(t, 123, mockedService.TestData().Get("something").Data())
-	}
-}
-
 func Test_Mock_On(t *testing.T) {
 	t.Parallel()
